just using uint32 for masking buttons
git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@1983 fa19cda8-b258-4bce-aa1a-843e42da0b19

+++ src/chan_sccp.h
@@ -819,7 +819,7 @@ struct sccp_device {
 	unsigned int				mwilamp: 3;				/*!< MWI/Lamp to indicate MailBox Messages */
 	unsigned int				mwioncall: 1;				/*!< MWI On Call Support (Boolean, default=on) */
 	unsigned int				softkeysupport: 1;			/*!< Soft Key Support (Boolean, default=on) */
-	uint64_t				mwilight;				/*!< MWI/Light bit field to to store mwi light for each line and device (offset 0 is current device state) */
+	uint32_t				mwilight;				/*!< MWI/Light bit field to to store mwi light for each line and device (offset 0 is current device state) */
 
 	unsigned int				transfer: 1;				/*!< Transfer Support (Boolean, default=on) */
 	//unsigned int				conference: 1;				/*!< Conference Support (Boolean, default=on) */
